pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.2.1",
    "fuzzywuzzy[speedup]>=0.18.0",
    "loguru>=0.7.3",
    "pipecat-ai[local,silero]>=0.0.98",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
    "tavily-python>=0.7.19",
    "websockets>=16.0",
]
